feat(devops): implement CI pipeline for monorepo validation
