+++ .github/workflows/ci.yml
@@ -7,24 +7,132 @@ on:
     branches: ["main"]
 
 jobs:
-  build:
-    name: Build and Test
+  web:
+    name: Web
     runs-on: ubuntu-latest
-
     steps:
-      - name: Checkout code
+      - name: Checkout repository
         uses: actions/checkout@v4
-        with:
-          fetch-depth: 2
 
-      - name: Setup Node.js
+      - name: Setup Node
         uses: actions/setup-node@v4
         with:
           node-version: 20
-          cache: 'npm'
+          cache: "npm"
+
+      - name: Restore Turbo cache
+        uses: actions/cache@v4
+        with:
+          path: .turbo
+          key: ${{ runner.os }}-turbo-web-${{ hashFiles('package-lock.json', 'turbo.json') }}
+          restore-keys: |
+            ${{ runner.os }}-turbo-web-
 
       - name: Install dependencies
-        run: npm install
+        run: npm ci
 
-      - name: Build Monorepo
-        run: npm run build
+      - name: Lint web
+        run: npm run lint -w web
+
+      - name: Build web
+        run: npm run build -w web -- --webpack
+
+      - name: Typecheck web
+        run: npm run typecheck -w web
+
+  packages:
+    name: Package Validation (${{ matrix.workspace }})
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        workspace: ["@devconsole/ui", "@devconsole/soroban-utils"]
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Setup Node
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: "npm"
+
+      - name: Restore Turbo cache
+        uses: actions/cache@v4
+        with:
+          path: .turbo
+          key: ${{ runner.os }}-turbo-packages-${{ matrix.workspace }}-${{ hashFiles('package-lock.json', 'turbo.json') }}
+          restore-keys: |
+            ${{ runner.os }}-turbo-packages-
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Validate package
+        run: npx turbo run lint typecheck build --filter=${{ matrix.workspace }}
+
+  api:
+    name: API
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Setup Node
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: "npm"
+
+      - name: Restore Turbo cache
+        uses: actions/cache@v4
+        with:
+          path: .turbo
+          key: ${{ runner.os }}-turbo-api-${{ hashFiles('package-lock.json', 'turbo.json') }}
+          restore-keys: |
+            ${{ runner.os }}-turbo-api-
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Generate Prisma client
+        run: npm run prisma:generate -w api
+
+      - name: Lint API
+        run: npm run lint -w api
+
+      - name: Build API
+        run: npm run build -w api
+
+      - name: Test API
+        run: npm run test -w api
+
+  contracts:
+    name: Contracts (${{ matrix.contract }})
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        contract: ["auth-tester", "error-trigger", "source-registry", "types-tester"]
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Setup Rust
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: wasm32-unknown-unknown
+
+      - name: Restore Cargo cache
+        uses: Swatinem/rust-cache@v2
+        with:
+          workspaces: contracts/${{ matrix.contract }}
+
+      - name: Test contract
+        run: cargo test --manifest-path contracts/${{ matrix.contract }}/Cargo.toml
+
+      - name: Build contract
+        run: cargo build --release --manifest-path contracts/${{ matrix.contract }}/Cargo.toml
